re-org project items
