+++ Source/nHydrate.Core.SQLGeneration/SQLEmit.cs
@@ -2,6 +2,7 @@
 using nHydrate.Generator.Common;
 using nHydrate.Generator.Common.Util;
 using nHydrate.Generator.Models;
+using nHydrate.Generator.Util;
 using System;
 using System.Collections.Generic;
 using System.Data;
