+++ Source/nHydrate.DslPackage/Forms/CueBox.cs
@@ -1,0 +1,38 @@
+using System;
+using System.Runtime.InteropServices;
+
+namespace nHydrate.DslPackage.Forms
+{
+    public class CueTextBox : System.Windows.Forms.TextBox
+    {
+        private string _cue;
+
+        public string Cue
+        {
+            get { return _cue; }
+            set
+            {
+                _cue = value;
+                UpdateCue();
+            }
+        }
+
+        private void UpdateCue()
+        {
+            if (this.IsHandleCreated && _cue != null)
+            {
+                SendMessage(this.Handle, 0x1501, (IntPtr)1, _cue);
+            }
+        }
+
+        protected override void OnHandleCreated(System.EventArgs e)
+        {
+            base.OnHandleCreated(e);
+            UpdateCue();
+        }
+
+        // P/Invoke
+        [DllImport("user32.dll", CharSet = CharSet.Unicode)]
+        private static extern IntPtr SendMessage(IntPtr hWnd, int msg, IntPtr wp, string lp);
+    }
+}
