+++ Source/nHydrate.DslPackage/Forms/RegistrationForm.cs
@@ -1,7 +1,7 @@
 #pragma warning disable 0168
 using nHydrate.DslPackage.Objects;
-using nHydrate.Generator.Common.GeneratorFramework;
 using nHydrate.Generator.Common.Util;
+using nHydrate.Generator.GeneratorFramework;
 using nHydrate.ServerObjects;
 using System;
 using System.Windows.Forms;
