+++ Source/nHydrate.DslPackage/Objects/ModelGenerateDomain.cs
@@ -5,6 +5,7 @@ using nHydrate.Generator.Common.GeneratorFramework;
 using nHydrate.Generator.Common.Util;
 using nHydrate.Generator.Models;
 using nHydrate.Generator.ModelUI;
+using nHydrate.Generator.Util;
 using System;
 using System.Collections.Generic;
 using System.IO;
