--- Source/nHydrate.Generator.Common/Util/GlobalHelper.cs
@@ -1,23 +1,0 @@
-using nHydrate.Generator.Common.Logging;
-using System;
-
-namespace nHydrate.Generator.Common.Util
-{
-    public static class GlobalHelper
-    {
-        public static void ShowError(Exception ex)
-        {
-            ShowError(ex.ToString());
-        }
-
-        public static void ShowError(string message)
-        {
-            nHydrateLog.LogError(new Exception(message));
-            var F = new nHydrate.Generator.Common.Forms.ErrorForm("An error has occurred!", message);
-            F.ShowDialog();
-        }
-
-    }
-
-}
-
