+++ Source/nHydrate.Generator.Common/Util/ProjectItemCacheItem.cs
@@ -1,6 +1,6 @@
 namespace nHydrate.Generator.Common.Util
 {
-    internal class ProjectItemCacheItem
+    public class ProjectItemCacheItem
     {
         public string Name { get; set; }
         public string Dependent { get; set; }
