+++ Source/nHydrate.Generator.EFCodeFirstNetCore/Generators/Entity/EntityGeneratedTemplate.cs
@@ -1,5 +1,6 @@
 using nHydrate.Generator.Common.Util;
 using nHydrate.Generator.Models;
+using nHydrate.Generator.Util;
 using System;
 using System.Linq;
 using System.Text;
