+++ Source/nHydrate.Generator.EFCodeFirstNetCore/Generators/ViewEntity/ViewEntityGeneratedTemplate.cs
@@ -1,6 +1,7 @@
 #pragma warning disable 0168
 using nHydrate.Generator.Common.Util;
 using nHydrate.Generator.Models;
+using nHydrate.Generator.Util;
 using System.Linq;
 using System.Text;
