+++ Source/nHydrate.Generator.PostgresInstaller/SQLHelper.cs
@@ -3,6 +3,7 @@ using nHydrate.Generator.Common;
 using nHydrate.Generator.Common.Util;
 using nHydrate.Generator.Models;
 using nHydrate.Generator.PostgresInstaller.ProjectItemGenerators;
+using nHydrate.Generator.Util;
 using System;
 using System.Collections.Generic;
 using System.Data;
