+++ Source/nHydrate.Generator/Models/RowEntry.cs
@@ -1,6 +1,7 @@
 #pragma warning disable 0168
 using nHydrate.Generator.Common;
 using nHydrate.Generator.Common.Util;
+using nHydrate.Generator.Util;
 using System;
 using System.Xml;
