+++ Source/nHydrate.Generator/ProjectItemGenerators/BaseProjectGenerator.cs
@@ -1,7 +1,8 @@
 #pragma warning disable 0168
 using nHydrate.Generator.Common.GeneratorFramework;
-using nHydrate.Generator.Common.Util;
+using nHydrate.Generator.GeneratorFramework;
 using nHydrate.Generator.Models;
+using nHydrate.Generator.Util;
 using System;
 using System.IO;
